removed bugs from backend
