+++ server/src/controller/interfaces.js
@@ -1,0 +1,11 @@
+const Example = require('./models/example');
+
+exports.apiData = (req, res) => {
+  name: string;
+  age: number;
+  pass: boolean;
+}
+
+exports.notFoundData = (req, res) => {
+  message: string;
+}
