+++ server/src/middleware/logger.js
@@ -1,0 +1,5 @@
+const lodger = ( req, res, next) => {
+    console.log(`Endpoint which has been hit: ${req.url}`);
+}
+
+export default logger;
