--- server/src/routes/root.ts
@@ -1,17 +1,0 @@
-import express, { Request, Response, Router } from "express";
-
-import { apiData } from "../types/interfaces";
-
-const dataToSend: apiData = {
-  name: "John Doe",
-  age: 13,
-  pass: true,
-};
-
-const router: Router = express.Router();
-
-router.get("/", (req: Request, res: Response) => {
-  res.send(dataToSend);
-});
-
-export default router;
