--- server/src/types/interfaces.ts
@@ -1,9 +1,0 @@
-export interface apiData {
-  name: string;
-  age: number;
-  pass: boolean;
-}
-
-export interface notFoundData {
-  message: string;
-}
